Refactor build script into multiple files
