+++ packages/default-plugins/commands/buildAll.ts
@@ -1,0 +1,7 @@
+import buildDefaultPlugins from '../buildDefaultPlugins';
+
+const buildAll = (outputDirectory: string) => {
+	return buildDefaultPlugins(outputDirectory, async () => { });
+};
+
+export default buildAll;
